creating some wrapper module
