+++ braindecode/modules/modules.py
@@ -30,37 +30,6 @@ class Chomp1d(nn.Module):
         return x[:, :, : -self.chomp_size].contiguous()
 
 
-class Expression(nn.Module):
-    """Compute given expression on forward pass.
-
-    Parameters
-    ----------
-    expression_fn : callable
-        Should accept variable number of objects of type
-        `torch.autograd.Variable` to compute its output.
-    """
-
-    def __init__(self, expression_fn):
-        super(Expression, self).__init__()
-        self.expression_fn = expression_fn
-
-    def forward(self, *x):
-        return self.expression_fn(*x)
-
-    def __repr__(self):
-        if hasattr(self.expression_fn, "func") and hasattr(
-            self.expression_fn, "kwargs"
-        ):
-            expression_str = "{:s} {:s}".format(
-                self.expression_fn.func.__name__, str(self.expression_fn.kwargs)
-            )
-        elif hasattr(self.expression_fn, "__name__"):
-            expression_str = self.expression_fn.__name__
-        else:
-            expression_str = repr(self.expression_fn)
-        return self.__class__.__name__ + "(expression=%s) " % expression_str
-
-
 class InceptionBlock(nn.Module):
     """
     Inception block module.
@@ -81,46 +50,6 @@ class InceptionBlock(nn.Module):
 
     def forward(self, x):
         return torch.cat([branch(x) for branch in self.branches], 1)
-
-
-class IntermediateOutputWrapper(nn.Module):
-    """Wraps network model such that outputs of intermediate layers can be returned.
-    forward() returns list of intermediate activations in a network during forward pass.
-
-    Parameters
-    ----------
-    to_select : list
-        list of module names for which activation should be returned
-    model : model object
-        network model
-
-    Examples
-    --------
-    >>> model = Deep4Net()
-    >>> select_modules = ['conv_spat','conv_2','conv_3','conv_4'] # Specify intermediate outputs
-    >>> model_pert = IntermediateOutputWrapper(select_modules,model) # Wrap model
-    """
-
-    def __init__(self, to_select, model):
-        if not len(list(model.children())) == len(list(model.named_children())):
-            raise Exception("All modules in model need to have names!")
-
-        super(IntermediateOutputWrapper, self).__init__()
-
-        modules_list = model.named_children()
-        for key, module in modules_list:
-            self.add_module(key, module)
-            self._modules[key].load_state_dict(module.state_dict())
-        self._to_select = to_select
-
-    def forward(self, x):
-        # Call modules individually and append activation to output if module is in to_select
-        o = []
-        for name, module in self._modules.items():
-            x = module(x)
-            if name in self._to_select:
-                o.append(x)
-        return o
 
 
 class TimeDistributed(nn.Module):
